% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oRiental_zodiac.R
\name{oRiental_zodiac}
\alias{oRiental_zodiac}
\title{Calculate oriental zodiac from year or date inclued year data}
\usage{
oRiental_zodiac(year, format = "\%Y/\%m/\%d")
}
\arguments{
\item{year}{A numeric data or date object format "year/month/day".}

\item{format}{Character data string. If you want to use a date object "m-d" or "Y/m/d",
change format.}
}
\description{
Calculate oriental zodiac from year or date inclued year data
}
\examples{
oRiental_zodiac(2020)
oRiental_zodiac("2020/12/25")

}
