% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRsign.R
\name{staRsign.ja}
\alias{staRsign.ja}
\title{Calculate the star sign in japanese from birthday}
\usage{
staRsign.ja(birthday, format = "\%m/\%d")
}
\arguments{
\item{birthday}{A date object in "month/day" or Objects of Class "Date" format of
Birthday whose want to calculate the star sign.}

\item{format}{Character date string. If you want to use a date object "m-d" or "Y/m/d",
you can change format.}
}
\description{
Calculate the star sign in japanese from birthday
}
\examples{
staRsign.ja("12/25")
staRsign.ja("2020/12/25", format = "\%Y/\%m/\%d")

}
