% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRsign.R
\name{staRsign}
\alias{staRsign}
\title{Calculate the star sign from birthday}
\usage{
staRsign(birthday, format = "\%m/\%d")
}
\arguments{
\item{birthday}{A date object in "month/dayya" format of Birthday whose want to calculate the
star sign.}

\item{format}{Character date string. If you want to use a date object "m-d" or "Y/m/d",
change format.}
}
\description{
Calculate the star sign from birthday
}
\examples{
staRsign("12/25")
staRsign("2020/12/25", format = "\%Y/\%m/\%d")

}
